Added two types of tests.

+++ src/CosmosSample/Benchmark.cs
@@ -6,17 +6,26 @@ namespace CosmosSample;
 public sealed class Benchmark : IAsyncDisposable
 {
     private const string DatabaseName = "index-experiments";
-    private const string ContainerName = "transtations";
 
     private readonly CosmosClient _client;
     private readonly TranslationRuleFactory _translationRuleFactory;
+    private readonly string _containerName;
+    private string _partitionKey;
 
     public Container? Container { get; private set; }
 
-    public Benchmark(CosmosClient client, int sourceSystemCount, int targetSystemsCount)
+    public Benchmark(CosmosClient client, int sourceSystemCount, int targetSystemsCount, bool useIdAsPartitionKey)
     {
         _client = client;
         _translationRuleFactory = new TranslationRuleFactory(sourceSystemCount, targetSystemsCount);
+
+        _partitionKey = useIdAsPartitionKey
+            ? "/id"
+            : "/SourceSystem";
+
+        _containerName = useIdAsPartitionKey
+            ? "transtationid"
+            : "transtationsystem";
     }
 
     public async ValueTask DisposeAsync()
@@ -27,39 +36,31 @@ public sealed class Benchmark : IAsyncDisposable
         }
     }
 
-    public async Task CreateContainerAsync()
+    public async Task CreateContainerAsync(bool deleteExisting)
     {
         DatabaseResponse databaseResponse = await _client.CreateDatabaseIfNotExistsAsync(
             id: DatabaseName,
             throughput: 400);
         Database database = databaseResponse.Database;
+
+        if (deleteExisting)
+        {
+            try
+            {
+                Container container0 = database.GetContainer(_containerName);
+                await container0.DeleteContainerAsync();
+                Console.WriteLine("Removed old container.");
+            }
+            catch (CosmosException)
+            {
+            }
+        }
 
         ContainerResponse containerResponse = await database.CreateContainerIfNotExistsAsync(
-            id: ContainerName,
-            partitionKeyPath: "/SourceSystem"
-            //partitionKeyPath: "/id"
+            id: _containerName,
+            partitionKeyPath: _partitionKey
         );
         Container = containerResponse.Container;
-    }
-
-    public async Task ResetContainerAsync()
-    {
-        DatabaseResponse databaseResponse = await _client.CreateDatabaseIfNotExistsAsync(
-            id: DatabaseName,
-            throughput: 400);
-        Database database = databaseResponse.Database;
-
-        try
-        {
-            Container container0 = database.GetContainer(ContainerName);
-            await container0.DeleteContainerAsync();
-            Console.WriteLine("Removed old container.");
-        }
-        catch (CosmosException)
-        {
-        }
-
-        await CreateContainerAsync();
     }
 
     public async Task SeedDataAsync(int count)
@@ -68,31 +69,85 @@ public sealed class Benchmark : IAsyncDisposable
         for (int i = 0; i < count; i++)
         {
             TranslationRule translationRule = _translationRuleFactory.Create();
-            if (i == 0)
-            {
-                translationRule.id = Guid.Empty;
-            }
+            //if (i == 0)
+            //{
+            //    translationRule.id = Guid.Empty;
+            //}
 
             ItemResponse<TranslationRule> itemResponse = await Container!.CreateItemAsync(translationRule);
             //await Container!.CreateItemAsync(
             //    translationRule,
             //    partitionKey: new PartitionKey("/SourceSystem"));
-            Console.WriteLine("Create item {0} {1} RUs", translationRule.id, itemResponse.RequestCharge);
+
+            Console.WriteLine("Create item #{0} {1} {2} RUs", i, translationRule.id, itemResponse.RequestCharge);
         }
 
         sw.Stop();
         Console.WriteLine("Seeded {0} items in {1}.", count, sw.Elapsed);
     }
+    
+    public async Task SeedDataAsync2(int count)
+    {
+        List<TranslationRule> translationRules = Enumerable.Range(1, count)
+            .Select(
+                i =>
+                {
+                    TranslationRule translationRule = _translationRuleFactory.Create();
+                    if (i == 0)
+                    {
+                        translationRule.id = Guid.Empty;
+                    }
+
+                    return translationRule;
+                })
+            .ToList();
+
+        var sw = Stopwatch.StartNew();
+
+        await Parallel.ForEachAsync(
+            translationRules,
+            async (x, ct) =>
+            {
+                ItemResponse<TranslationRule> itemResponse = await Container!.CreateItemAsync(x);
+                Console.WriteLine("Create item {0} {1} RUs", x.id, itemResponse.RequestCharge);
+            });
+
+        sw.Stop();
+
+        Console.WriteLine("Seeded {0} items in {1}.", count, sw.Elapsed);
+    }
 
     public async Task QueryAllDataAsync()
     {
-        using FeedIterator<TranslationRule> feed = Container!.GetItemQueryIterator<TranslationRule>(queryText: $"SELECT * FROM {ContainerName}");
+        using FeedIterator<TranslationRule> feed = Container!.GetItemQueryIterator<TranslationRule>(queryText: $"SELECT * FROM {_containerName}");
         while (feed.HasMoreResults)
         {
             FeedResponse<TranslationRule> response = await feed.ReadNextAsync();
             foreach (TranslationRule? item in response)
             {
-                Console.WriteLine($"Found item:\t{item.id}\t{response.RequestCharge} RUs");
+                Console.WriteLine($"Found item: {item.id} {item} {response.RequestCharge} RUs");
+                break;
+            }
+        }
+    }
+
+    public async Task QueryBySourceTargetDataAsync(string sourceSystem, string targetSystem)
+    {
+        QueryDefinition parameterizedQuery = new QueryDefinition(
+                query: $"SELECT * FROM {_containerName} p WHERE p.SourceSystem = @ss AND p.TargetSystem = @ts"
+            )
+            .WithParameter("@ss", sourceSystem)
+            .WithParameter("@ts", targetSystem);
+
+
+        using FeedIterator<TranslationRule> feed = Container!.GetItemQueryIterator<TranslationRule>(queryDefinition: parameterizedQuery);
+        while (feed.HasMoreResults)
+        {
+            FeedResponse<TranslationRule> response = await feed.ReadNextAsync();
+            foreach (TranslationRule? item in response)
+            {
+                Console.WriteLine($"Found item: {item.id} {item} {response.RequestCharge} RUs");
+                break;
             }
         }
     }
